requirements.txt
Crypto
python-whois
python-telegram-bot
pluginbase
python-nmap